CommonJS-ified *some* of the things
