+++ lib/sinon/extend.js
@@ -1,7 +1,5 @@
 "use strict";
 
-var sinon = require("./util/core");
-
 // Adapted from https://developer.mozilla.org/en/docs/ECMAScript_DontEnum_attribute#JScript_DontEnum_Bug
 var hasDontEnumBug = (function () {
     var obj = {
@@ -54,13 +52,12 @@ var hasDontEnumBug = (function () {
  *
  * Returns the extended target
  */
-function extend(target /*, sources */) {
+module.exports = function extend(target /*, sources */) {
     var sources = Array.prototype.slice.call(arguments, 1);
-    var source, i, prop;
+    var i = 0;
+    var source, prop;
 
-    for (i = 0; i < sources.length; i++) {
-        source = sources[i];
-
+    while ((source = sources[i++])) {
         for (prop in source) {
             if (source.hasOwnProperty(prop)) {
                 target[prop] = source[prop];
@@ -75,6 +72,4 @@ function extend(target /*, sources */) {
     }
 
     return target;
-}
-
-sinon.extend = extend;
+};
