+++ lib/sinon/match.js
@@ -8,12 +8,13 @@
  */
 "use strict";
 
-
-require("./typeOf");
-var sinon = require("./util/core");
+var create = require("./util/core/create");
+var deepEqual = require("./util/core/deep-equal");
+var functionName = require("./util/core/function-name");
+var typeOf = require("./typeOf");
 
 function assertType(value, type, name) {
-    var actual = sinon.typeOf(value);
+    var actual = typeOf(value);
     if (actual !== type) {
         throw new TypeError("Expected type of " + name + " to be " +
             type + ", but was " + actual);
@@ -42,11 +43,11 @@ function matchObject(expectation, actual) {
                 if (!exp.test(act)) {
                     return false;
                 }
-            } else if (sinon.typeOf(exp) === "object") {
+            } else if (typeOf(exp) === "object") {
                 if (!matchObject(exp, act)) {
                     return false;
                 }
-            } else if (!sinon.deepEqual(exp, act)) {
+            } else if (!deepEqual(exp, act)) {
                 return false;
             }
         }
@@ -55,15 +56,15 @@ function matchObject(expectation, actual) {
 }
 
 function match(expectation, message) {
-    var m = sinon.create(matcher);
-    var type = sinon.typeOf(expectation);
+    var m = create(matcher);
+    var type = typeOf(expectation);
     switch (type) {
     case "object":
         if (typeof expectation.test === "function") {
             m.test = function (actual) {
                 return expectation.test(actual) === true;
             };
-            m.message = "match(" + sinon.functionName(expectation.test) + ")";
+            m.message = "match(" + functionName(expectation.test) + ")";
             return m;
         }
         var str = [];
@@ -105,12 +106,12 @@ function match(expectation, message) {
         if (message) {
             m.message = message;
         } else {
-            m.message = "match(" + sinon.functionName(expectation) + ")";
+            m.message = "match(" + functionName(expectation) + ")";
         }
         break;
     default:
         m.test = function (actual) {
-            return sinon.deepEqual(expectation, actual);
+            return deepEqual(expectation, actual);
         };
     }
     if (!m.message) {
@@ -126,7 +127,7 @@ matcher.or = function (m2) {
         m2 = match(m2);
     }
     var m1 = this;
-    var or = sinon.create(matcher);
+    var or = create(matcher);
     or.test = function (actual) {
         return m1.test(actual) || m2.test(actual);
     };
@@ -141,7 +142,7 @@ matcher.and = function (m2) {
         m2 = match(m2);
     }
     var m1 = this;
-    var and = sinon.create(matcher);
+    var and = create(matcher);
     and.test = function (actual) {
         return m1.test(actual) && m2.test(actual);
     };
@@ -176,7 +177,7 @@ match.same = function (expectation) {
 match.typeOf = function (type) {
     assertType(type, "string", "type");
     return match(function (actual) {
-        return sinon.typeOf(actual) === type;
+        return typeOf(actual) === type;
     }, "typeOf(\"" + type + "\")");
 };
 
@@ -184,7 +185,7 @@ match.instanceOf = function (type) {
     assertType(type, "function", "type");
     return match(function (actual) {
         return actual instanceof type;
-    }, "instanceOf(" + sinon.functionName(type) + ")");
+    }, "instanceOf(" + functionName(type) + ")");
 };
 
 function createPropertyMatcher(propertyTest, messagePrefix) {
@@ -201,7 +202,7 @@ function createPropertyMatcher(propertyTest, messagePrefix) {
                     !propertyTest(actual, property)) {
                 return false;
             }
-            return onlyProperty || sinon.deepEqual(value, actual[property]);
+            return onlyProperty || deepEqual(value, actual[property]);
         }, message);
     };
 }
@@ -226,4 +227,4 @@ match.array = match.typeOf("array");
 match.regexp = match.typeOf("regexp");
 match.date = match.typeOf("date");
 
-sinon.match = match;
+module.exports = match;
