+++ lib/sinon/stub.js
@@ -176,3 +176,10 @@ for (var method in sinon.behavior) {
 
 sinon.extend(stub, proto);
 sinon.stub = stub;
+sinon.createStubInstance = function (constructor) {
+    if (typeof constructor !== "function") {
+        throw new TypeError("The constructor should be a function.");
+    }
+    return sinon.stub(sinon.create(constructor.prototype));
+};
+
