+++ lib/sinon/typeOf.js
@@ -8,9 +8,7 @@
  */
 "use strict";
 
-var sinon = require("./util/core");
-
-function typeOf(value) {
+module.exports = function typeOf(value) {
     if (value === null) {
         return "null";
     } else if (value === undefined) {
@@ -18,6 +16,4 @@ function typeOf(value) {
     }
     var string = Object.prototype.toString.call(value);
     return string.substring(8, string.length - 1).toLowerCase();
-}
-
-sinon.typeOf = typeOf;
+};
