+++ lib/sinon/util/core/create.js
@@ -1,0 +1,9 @@
+"use strict";
+
+var Klass = function () {};
+
+module.exports = function create(proto) {
+    Klass.prototype = proto;
+    return new Klass();
+};
+
