+++ lib/sinon/util/core/default-config.js
@@ -1,0 +1,10 @@
+"use strict";
+
+module.exports = {
+    injectIntoThis: true,
+    injectInto: null,
+    properties: ["spy", "stub", "mock", "clock", "server", "requests"],
+    useFakeTimers: true,
+    useFakeServer: true
+};
+
