+++ lib/sinon/util/core/times-in-words.js
@@ -1,0 +1,8 @@
+"use strict";
+
+var array = [null, "once", "twice", "thrice"];
+
+module.exports = function timesInWords(count) {
+    return array[count] || (count || 0) + " times";
+};
+
